refactor(apollo): rename files components
