+++ client/apollo/react/src/Svg/Svg.tsx
