+++ client/apollo/react/src/index.ts
@@ -1,6 +1,6 @@
 import "@axa-fr/design-system-apollo-css/dist/common/tokens.scss";
 import "@fontsource/source-sans-pro";
 
-export { Button, buttonVariants, type ButtonVariants } from "./Button";
+export { Button, buttonVariants, type ButtonVariants } from "./Button/Button";
 export { renderButton, renderButtonAll } from "./Button/render";
-export { Svg } from "./Svg";
+export { Svg } from "./Svg/Svg";
